Remove whitespace as potential singular start character, allow for any.
Should help with detecting python examples in docstrings which start
with ">>>".  They do not have a comment character at the start of their
lines though and shouldn't be detected.

I don't know why white space is a possible start character on its own.
To allow for indentation of table-of-comments, it should be
\s* + comment_character for each comment_character.

That's what I've done here.